test for math.
test for math notation in markdown.
